Added UserMapper, changed methods in UserController

+++ src/main/java/com/women/online/shop/controllers/UserController.java
@@ -1,11 +1,13 @@
 package com.women.online.shop.controllers;
 
 import com.women.online.shop.controllers.dto.UserDto;
+import com.women.online.shop.controllers.mappers.UserMapper;
 import com.women.online.shop.entities.Address;
 import com.women.online.shop.entities.Name;
 import com.women.online.shop.entities.User;
 import com.women.online.shop.repositoties.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -23,36 +25,36 @@ public class UserController {
     }
 
     @PostMapping("/users")
-    public void create(@RequestBody UserDto userDto) {
-        final User user =
-                new User(
-                        new Name(userDto.getFirstName(), null, userDto.getLastName()),
-                        userDto.getEmail(),
-                        new Address(userDto.getCountry(), userDto.getCity(), userDto.getStreet(),
-                                userDto.getHomeNumber(), userDto.getApartment())
-                );
+    public ResponseEntity<User> create(@RequestBody UserDto userDto) {
+        User user = UserMapper.convertToEntity(userDto);
         userRepository.save(user);
+        return new ResponseEntity<>(user, HttpStatus.CREATED);
     }
 
     @GetMapping("/users")
     public ResponseEntity<List<User>> getUsers() {
-        final List<User> all = userRepository.findAll();
-        return ResponseEntity.ok(all);
+        final List<User> users = userRepository.findAll();
+        return ResponseEntity.ok(users);
     }
 
     @GetMapping("/users/{id}")
     public ResponseEntity<Optional<User>> getUser(@PathVariable String id) {
         Optional<User> user = userRepository.findById(id);
-        return ResponseEntity.ok(user);
+        if (user.isPresent()) {
+            return ResponseEntity.ok(user);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/users/{id}")
-    public void delete(@PathVariable String id) {
-        userRepository.deleteById(id);
-    }
-
-    @PostConstruct
-    public void postConstract() {
-        System.out.println("PostConstruct");
+    public ResponseEntity<User> delete(@PathVariable String id) {
+        Optional<User> user = userRepository.findById(id);
+        if (user.isPresent()) {
+            userRepository.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
